// apps/admin/src/generated/api/schemas/listRolesParams.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Common Admin API
 * API for the common admin starter template
 * OpenAPI spec version: 0.1.0
 */
import type { ListRolesStatus } from "./listRolesStatus";
import type { Object } from "./object";

export type ListRolesParams = {
  /**
   * @minimum 1
   */
  page?: Object;
  /**
   * @minimum 1
   * @maximum 100
   */
  pageSize?: Object;
  search?: string;
  sort?: string;
  status?: ListRolesStatus;
};